Fix tests to invite new back office user via email
