+++ features/back_office/create_registration.feature
@@ -4,6 +4,7 @@ Feature: Back office user completes assisted digital registration for a user
   I need to register a waste exemption activity
   So that I can register exemptions on behalf of assisted digital users
 
+@email
   Scenario: Registration of a limited company by a back office user
   	Given I sign in as an admin agent
      When I complete a "limited_company" registration
