+++ features/support/email_helpers.rb
@@ -16,19 +16,19 @@ def email_exists?(app, registration, expected_text)
 
   # Load the appropriate front or back office Notify page:
   if app == :bo
-    visit(Quke::Quke.config.custom["urls"]["back_office_email"])
+    visit(Quke::Quke.config.custom["urls"]["back_office_notify"])
   else
-    visit(Quke::Quke.config.custom["urls"]["front_office_email"])
+    visit(Quke::Quke.config.custom["urls"]["front_office_notify"])
   end
 
   # We don't know whether the applicant or contact email will be sent first, so try both.
   # Try the applicant email:
   expected_text_for_applicant = expected_text + [registration[:applicant][:email]]
-  return true if @world.journey.last_notify_msg_page.email_has_text?(expected_text_for_applicant)
+  return true if @world.journey.last_message_page.email_has_text?(expected_text_for_applicant)
 
   # If that doesn't work, try the contact email:
   expected_text_for_contact = expected_text + [registration[:contact][:email]]
-  return true if @world.journey.last_notify_msg_page.email_has_text?(expected_text_for_contact)
+  return true if @world.journey.last_message_page.email_has_text?(expected_text_for_contact)
 
   puts "Email not found"
   false
